requirements.txt
streamlit
pandas
scikit-learn
joblib
numpy
python-dotenv
llamaapi